Rename solver submodules
This commit renames the classic and sharpclaw solver modules to be
solver removing the awkward double import.  This is a redo of another
commit that did this same change but was not able to be merged in
automatically and easily by hand.
