+++ src/petclaw/__init__.py
@@ -22,8 +22,8 @@ from .state import State
 from .cfl import CFL
 
 __all__.extend(['ClawSolver1D','ClawSolver2D','ClawSolver3D','SharpClawSolver1D','SharpClawSolver2D'])
-from .classic.clawpack import ClawSolver1D,ClawSolver2D,ClawSolver3D
-from .sharpclaw.sharpclaw import SharpClawSolver1D,SharpClawSolver2D
+from .classic.solver import ClawSolver1D,ClawSolver2D,ClawSolver3D
+from .sharpclaw.solver import SharpClawSolver1D,SharpClawSolver2D
 
 __all__.append('BC')
 from clawpack.pyclaw.solver import BC
