+++ src/petclaw/classic/solver.py
@@ -7,16 +7,16 @@ determine the containing claw_package to use.
 """
 
 from __future__ import absolute_import
-from clawpack.pyclaw.classic import clawpack
+from clawpack.pyclaw.classic import solver
 
-class ClawSolver1D(clawpack.ClawSolver1D):
+class ClawSolver1D(solver.ClawSolver1D):
     r"""
     PetClaw solver for 1D problems using classic Clawpack algorithms.
 
     This class implements nothing; it just inherits from ClawSolver1D.
     """
 
-class ClawSolver2D(clawpack.ClawSolver2D):
+class ClawSolver2D(solver.ClawSolver2D):
     r"""
     PetClaw solver for 2D problems using classic Clawpack algorithms.
 
@@ -25,7 +25,7 @@ class ClawSolver2D(clawpack.ClawSolver2D):
     Note that only the fortran routines are supported for now in 2D.
     """
     
-class ClawSolver3D(clawpack.ClawSolver3D):
+class ClawSolver3D(solver.ClawSolver3D):
     r"""
     PetClaw solver for 3D problems using classic Clawpack algorithms.
