+++ src/pyclaw/classic/solver.py
